Add files via upload
Accomplished data scientist with 5 years of experience leveraging statistical models, machine learning algorithms, and predictive analytics to extract insights and optimize business solutions. Strong math and statistical knowledge with a proven ability to communicate complex findings to stakeholders in an accessible manner. Seeking to leverage analytical abilities and passion for data-driven decision making to improve products and processes as a data scientist